mars2mars: Reorganise rules
